Create README.md
ABOUT ME
Tambol Nelson is a young boy in Ghana who is a graphics designer, a farmer and a student with a  highly skilled professional and specialization in graphics design and agriculture. He focuses on encouraging the youth to creates jobs for themselves particularly through agriculture. With a strong passion for technology and innovation, Nelson embarked on journey to bridge the gap between between creativity and technology. With his technical expertise and an eye for design, he established his own graphics design site, showing his remarkable skills and offerings graphics design services to clients worldwide. His unwavering commitment to innovation in the digital world makes him a sought-after professional in the field of graphics design. He is also a talented and driven individual who is making a significant impact in the tech industry and agriculture sector. He is an inspiration to young people everywhere, and is proof that anything is possible if you set your mind to it. I'm on a strong believe that we can develop Africa, particularly Ghana by investing to agriculture, hence encouraging the youth to get in involved. As the saying goes "people become great whilst others were sleeping".